Made projects section modular along with a few wording changes.
