+++ src/Components/Contact2.js
@@ -1,9 +1,52 @@
-import React from 'react'
-import emailjs from 'emailjs-com'
-import './Contact2.scss'
-import Button from 'react-bootstrap/Button'
+import React from 'react';
+import emailjs from 'emailjs-com';
+import './Contact2.scss';
+import Button from 'react-bootstrap/Button';
+import { motion, useAnimation } from 'framer-motion';
+import { useInView } from "react-intersection-observer";
+import { useEffect } from 'react';
+
+const contactVariants = {
+    hidden: {
+        opacity: 0,
+        y: '30'
+    },
+    visible: {
+        opacity: 1,
+        y: 0,
+        transition: {
+            delay: .3,
+            duration: .75,
+            ease: 'easeOut'
+        }
+    }
+}
+
+// const picVariants = {
+//     hidden: {
+//         opacity: 0,
+//         y: '30'
+//     },
+//     visible: {
+//         zIndex: 1,
+//         opacity: 1,
+//         y: 0,
+//         transition: {
+//             duration: .75,
+//             ease: 'easeOut'
+//         }
+//     }
+// }
 
 const Contact2 = () => {
+    const controls = useAnimation();
+    const [ref, inView] = useInView();
+    useEffect(() => {
+        if (inView) {
+            controls.start("visible");
+        }
+    }, [controls, inView]);
+
     function sendEmail(e) {
         e.preventDefault();
 
@@ -22,25 +65,30 @@ const Contact2 = () => {
 
     return (
         <div className="contact-section2">
-            <div className='form-container2'>
+            <motion.div className='form-container2'
+                variants={contactVariants}
+                animate={controls}
+                initial="hidden"
+                ref={ref}
+            >
                 <div className='contact-info'>
                     <h2>Contact Me</h2>
                     <p>(212)951-1427</p>
-                    <p>dwainegnd@gmail.com</p>
-                    <p>github.com/dwainejade</p>
+                    <a href="mailto:dwainegnd@gmail.com">dwainegnd@gmail.com</a>
+                    <a href="https://github.com/dwainejade">github.com/dwainejade</a>
                 </div>
 
                 <form onSubmit={sendEmail} id='email-form'>
                     <h2>Get in Touch</h2>
-                    <input type="text" name='name' placeholder='Your name' required />
-                    <input type="email" name='email' placeholder='Your email' required />
+                    <input type="text" name='name' placeholder='Name' required />
+                    <input type="email" name='email' placeholder='Email' required />
                     <input type="text" name='subject' placeholder='Subject' required />
-                    <textarea id="message" name='message' placeholder='Type your message here...' required />
+                    <textarea id="message" name='message' placeholder='Message...' required />
                     <Button variant="primary" type="submit">
                         SEND
                     </Button>
                 </form>
-            </div>
+            </motion.div>
         </div>
     )
 }
