+++ src/Components/Hero.js
@@ -45,7 +45,7 @@ function Hero() {
             >
                 <h1>Dwaine Matthew</h1>
                 <h2>Web Developer</h2>
-                <p>Full-stack developer specialized in frontend and backend development for complex scalable web apps.</p>
+                <p>Full-stack developer specialized in frontend design. Currently having too much fun with React.</p>
             </motion.div>
             <motion.div className="hero-pic"
                 variants={picVariants}
@@ -54,8 +54,6 @@ function Hero() {
             >
                 <img alt="profile-pic" src={profile_pic} />
             </motion.div>
-
-
         </div>
     )
 }
