+++ src/Components/Skills.scss
@@ -4,7 +4,7 @@
   justify-content: center;
   align-items: center;
   padding: 2rem;
-  height: 100vh;
+  // height: 100vh;
   color: white;
 
   .divider {
